make the frame argument a long to prevent an easy crash

+++ WallyMapEditor/src/Rendering/RaylibCanvas.cs
@@ -119,7 +119,7 @@ public class RaylibCanvas : ICanvas
         ), (int)priority);
     }
 
-    public void DrawAnim(Gfx gfx, string animName, int frame, WmsTransform trans, DrawPriorityEnum priority, object? caller, uint? loopLimit = null)
+    public void DrawAnim(Gfx gfx, string animName, long frame, WmsTransform trans, DrawPriorityEnum priority, object? caller, uint? loopLimit = null)
     {
         Animator.DrawAnim(gfx, animName, frame, trans, priority, caller, loopLimit);
     }
